+++ rencfsMultiplatform/src/commonMain/kotlin/rs/xor/rencfs/krencfs/ui/screens/SettingsScreen.kt
@@ -72,62 +72,3 @@ private fun SettingItem(
         )
     }
 }
-
-@Composable
-fun AboutScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        Text(
-            text = "About Rencfs",
-            style = MaterialTheme.typography.headlineMedium,
-            modifier = Modifier.padding(bottom = 16.dp)
-        )
-
-        Card(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column(modifier = Modifier.padding(16.dp)) {
-                InfoItem(
-                    title = "Version",
-                    content = "1.0.0"
-                )
-                InfoItem(
-                    title = "Description",
-                    content = "Rencfs is an encrypted filesystem implementation in Kotlin"
-                )
-                InfoItem(
-                    title = "License",
-                    content = "MIT License"
-                )
-                InfoItem(
-                    title = "Repository",
-                    content = "github.com/radumarias/rencfs-kotlin"
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun InfoItem(
-    title: String,
-    content: String
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp)
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleMedium
-        )
-        Text(
-            text = content,
-            style = MaterialTheme.typography.bodyMedium
-        )
-    }
-}
